+++ packages/jsdoc-plugins/lib/statics-augmentor/augmentstatics.js
